+++ at/fov/raycast.h
@@ -5,11 +5,13 @@
 #ifndef AT_RAYCAST_H
 #define AT_RAYCAST_H
 
+#include <stddef.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-void at_do_raycast_fov(int *view, int *grid, unsigned w, unsigned h, double r,
+void at_do_raycast_fov(int *view, int *grid, size_t w, size_t h, double r,
         int cx, int cy);
 
 #ifdef __cplusplus
