+++ at/fov/shadowcast.h
@@ -5,12 +5,14 @@
 #ifndef AT_SHADOWCAST_H
 #define AT_SHADOWCAST_H
 
+#include <stddef.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-void at_do_shadowcast_fov(int *view, int *grid, unsigned w, unsigned h,
-        double r, int cx, int cy);
+void at_do_shadowcast_fov(int *view, int *grid, size_t w, size_t h, double r,
+        int cx, int cy);
 
 #ifdef __cplusplus
 }
