+++ at/path/line.c
@@ -9,7 +9,7 @@
 
 
 
-void at_path_line(int *path, unsigned *sz, int x0, int y0, int x1, int y1)
+void at_path_line(int *path, size_t *sz, int x0, int y0, int x1, int y1)
 {
         int dx = abs(x1 - x0);
         int sx = x0 < x1 ? 1 : -1;
